requirements.txt
# LLM & Agents (Compatible versions)
openai==1.54.4
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.45
langchain-community==0.3.5

# Vector Store
chromadb==0.5.20

# Document Processing
python-pptx==1.0.2
pypdf==5.1.0

# Web Frameworks
streamlit==1.40.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Data Processing
pandas==2.2.3
pyyaml==6.0.2
python-dotenv==1.0.1

# Utilities
requests==2.32.3
Pillow==11.0.0
httpx==0.27.2
# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Code Quality
black==24.10.0
flake8==7.1.1